+++ src/forest.h
@@ -1,7 +1,7 @@
 /*
  * scrm is an implementation of the Sequential-Coalescent-with-Recombination Model.
  * 
- * Copyright (C) 2013, 2014 Paul R. Staab, Sha (Joe) Zhu and Gerton Lunter
+ * Copyright (C) 2013, 2014 Paul R. Staab, Sha (Joe) Zhu, Dirk Metzler and Gerton Lunter
  * 
  * This file is part of scrm.
  * 
@@ -36,6 +36,8 @@
 #ifndef scrm_src_forest
 #define scrm_src_forest
 
+#include <string>
+
 #include "macros.h" // Needs to be before cassert
 
 #include <vector>
@@ -44,6 +46,7 @@
 #include <cassert>
 #include <iostream> // ostreams
 #include <iomanip>  // Used for debug output
+#include <sstream>  // Used for debug output
 
 #include "contemporaries_container.h"
 #include "event.h"
@@ -63,6 +66,9 @@ enum eventCode { NOEVENT, EVENT, INIT_NULL};
 class Forest
 {
  public:
+  Node* readNewickNode( std::string &in_str, std::string::iterator &current_it, size_t parenthesis_balance = 0, Node* parent = NULL );
+  void readNewick(std::string &in_str);
+  ContemporariesContainer* contemporaries()  {return &this->contemporaries_;};
 
 #ifdef UNITTEST
   friend class TestForest;
@@ -74,10 +80,8 @@ class Forest
   friend class TimeInterval;
   friend class TimeIntervalIterator;
 
-  Forest();
   Forest(Model *model, RandomGenerator *random_generator);
   Forest(const Forest &current_forest);
-  Forest(Forest *current_forest);
   virtual ~Forest() {};
 
   //Getters & Setters
@@ -205,6 +209,8 @@ class Forest
   void printLocusSumStats(std::ostream &output) const;
   
  private:
+  Forest() { this->initialize(); }
+
   //Operations on the Tree
   Node* cut(const TreePoint &cut_point);