+++ src/summary_statistics/newick_tree.h
@@ -1,7 +1,7 @@
 /*
  * scrm is an implementation of the Sequential-Coalescent-with-Recombination Model.
  * 
- * Copyright (C) 2013, 2014 Paul R. Staab, Sha (Joe) Zhu and Gerton Lunter
+ * Copyright (C) 2013, 2014 Paul R. Staab, Sha (Joe) Zhu, Dirk Metzler and Gerton Lunter
  * 
  * This file is part of scrm.
  * 
@@ -31,21 +31,6 @@
 #include "summary_statistic.h"
 #include "../forest.h"
 
-/** Use the slower implementation below for std::to_string if compiled with 
- * '-DNTOSTRING'. This is useful for (cross) compiling with mingw, because it
- * misses support for std::to_string. */
-#ifdef NTOSTRING
-template <typename T> 
-std::string to_string(T value)
-{
-  std::ostringstream os ;
-  os << value ;
-  return os.str() ;
-}
-#else
-using std::to_string;
-#endif
-
 
 /**
  * @brief Save buffered tree along with the sequence position at which
@@ -59,7 +44,14 @@ struct NewickBuffer {
 class NewickTree : public SummaryStatistic
 {
  public:
-   NewickTree() {}
+   NewickTree() { 
+    precision_ = 6; 
+   }
+
+   NewickTree(size_t precision) { 
+    precision_ = precision; 
+   }
+
    //NewickTree(const NewickTree &nt) { (void)nt; }
    ~NewickTree() {}
 
@@ -81,6 +73,7 @@ class NewickTree : public SummaryStatistic
    std::string generateTree(Node *node, const Forest &forest,
                             const bool use_buffer = true);
    std::ostringstream output_buffer_;
+   size_t precision_;
 
    /**
     * A map to buffer already created subtrees indexed by their