+++ tests/unittests/test_forest.cc
@@ -260,7 +260,7 @@ class TestForest : public CppUnit::TestCase {
       CPPUNIT_ASSERT( event.isCoalescence() );
       count += (event.node() == forest2->active_node(0));
     };
-    CPPUNIT_ASSERT( 4950 < count && count < 5050 ); // ~5000
+    CPPUNIT_ASSERT( 4900 < count && count < 5100 ); // ~5000
 
     // Test with Pw Coalescence
     // active_node 0: Pop 1, 2 Contemporaries 
@@ -446,7 +446,6 @@ class TestForest : public CppUnit::TestCase {
 
     Event event;
     double tmp_event_time = 0.0;
-    size_t tmp_event_line = -1;
     for (size_t i = 0; i < 1000; ++i) {
       forest2.sampleEvent(*tii, tmp_event_time, event); 
       CPPUNIT_ASSERT( event.isNoEvent() || ( 0 <= event.time() && event.time() < forest2.nodes()->at(4)->height() ) );
@@ -700,31 +699,18 @@ class TestForest : public CppUnit::TestCase {
 
   void testCopyConstructor() {
     forest->createScaledExampleTree();
-
+    CPPUNIT_ASSERT( forest->coalescence_finished_ == true );
     Forest forest2 = Forest(*forest);
-    Forest forest3 = Forest(forest);
 
     CPPUNIT_ASSERT_EQUAL( forest->nodes()->size(), forest2.nodes()->size() );
-    CPPUNIT_ASSERT_EQUAL( forest->nodes()->size(), forest3.nodes()->size() );
-    /*
-    CPPUNIT_ASSERT_EQUAL( forest->secondary_roots_.size(), forest2.secondary_roots_.size() );
-    CPPUNIT_ASSERT_EQUAL( forest->secondary_roots_.size(), forest3.secondary_roots_.size() );
-    */
-
     CPPUNIT_ASSERT( forest2.model_ == forest->model_ );
-    CPPUNIT_ASSERT( forest3.model_ == forest->model_ );
 
     for (auto it = forest2.nodes()->iterator(); it.good(); ++it) {
       CPPUNIT_ASSERT( (*it)->label() <= 4 );
       CPPUNIT_ASSERT( (*it)->parent() != NULL || (*it)->first_child() != NULL );
     }
-    for (auto it = forest3.nodes()->iterator(); it.good(); ++it) {
-      CPPUNIT_ASSERT( (*it)->label() <= 4 );
-      CPPUNIT_ASSERT( (*it)->parent() != NULL || (*it)->first_child() != NULL );
-    }
 
     CPPUNIT_ASSERT( forest2.checkTree() );
-    CPPUNIT_ASSERT( forest3.checkTree() );
   }
 
   void testCheckForNodeAtHeight() {